reimplement compile.py
reimplement compile.py

+++ src/compile.py
@@ -1,61 +1,55 @@
 #!/usr/bin/python2.7
 '''compile scripts: this script expects to live at relative path:
-      src/compile/compile.py'''
+      src/compile.py'''
 import os
 import re
 import sys
 import copy
 import time
-from fl0w3r import error, exists, normpath, normdir
+from ansicolor import KYEL, KNRM
+from fl0w3r import error, exists, normpath, normdir, timestring, run, printw
 
-opt_flags, t, args = '-O3', time.localtime(), sys.argv
+time_string = timestring()
+opt_flags, args = '-O3', sys.argv
 bin_folder, source_folder = '', ''
 
 if os.name != "posix":
     error('defined for posix')
 
 head, tail = os.path.split(normdir(normpath(args[0])))
-# head, tail = os.path.split(head)
+source_folder = normpath(head)
+head, tail = os.path.split(head)
 
 if tail == 'bin':
     error('please run from src/ folder')
 else:
-    source_folder = head
-    if tail != 'compile':
-        error('expected to reside within a directory compile/')
-    head, tail = os.path.split(head)
     if tail != 'src':
-        error('expected compile/compile.py to reside within a directory src/');
+        error('expected compile.py to reside within a directory src/')
     if os.path.basename(__file__) != 'compile.py':
-        error('expected __file__ to be compile.py')   
-    bin_folder = head + '/bin/'
-
-if bin_folder == '':
-    error('bin/ folder undefined')
-
-if source_folder == '':
-    error('src/ folder undefined')
+        error('expected __file__ to be compile.py')
+    bin_folder = normpath(head + '/bin/')
+    if not exists(bin_folder):
+        run('mkdir -p ' + bin_folder)
+    bin_folder = normpath(bin_folder)
+    if not exists(bin_folder):
+        error('failed to create folder: ' + bin_folder)
 
 if not exists(source_folder):
     error('folder not found: ' + source_folder)
-
-if not exists(bin_folder):
-    error('folder not found: ' + bin_folder)
 
 if not os.path.isdir(bin_folder):
     error('not a directory: ' + bin_folder)
 
 if len(sys.argv) < 2:
     error('compile.py re-implemented 20170602 by ashlin richardson\n\tUsage: python src/compile.py src/name_of_program_to_compile')
-error('exit1')
 
-bin_folder = normpath(bin_folder)
+
 f = os.path.abspath(args[1].strip())
-cmdname = f.split("/")
-cmdname = cmdname[len(cmdname) - 1]
+cmdname = f.split("/")[-1]
+
 fileend = copy.deepcopy(cmdname)
-ext = fileend.split(".")
-ext = ext[len(ext) - 1]
+ext = fileend.split(".")[-1]
+
 compiler_cmd = ''
 if ext == 'cpp':
     compiler_cmd = 'g++ -w ' + opt_flags + ' -lm '
@@ -66,44 +60,34 @@ elif ext == 'py':
 else:
     error('file extension not supported')
 
-cmdname =  fileend.split('.')
-cmdname = cmdname[0]
-binname = bin_folder + cmdname
-cmd = 'rm -rf ' + binname
-print cmd
-a = os.system(cmd)
-command_args = ""
-time_string =( "%04d" % t.tm_year)+("%02d" % t.tm_mon)+("%02d" % t.tm_mday)+("%02d" % t.tm_hour)+("%02d" % t.tm_min)+("%02d" % t.tm_sec);
+cmdname =  fileend.split('.')[0]
+binname = normpath(normpath(bin_folder) + cmdname)
 
-if ext == 'cpp':
-    command_args = compiler_cmd + " "+f+ " -o "+binname+".exe"
+if exists(binname):
+    cmd = 'rm -rf ' + binname
+    print cmd
+    a = os.system(cmd)
+
+command_args, dat = '', None
+command_args = compiler_cmd + ' ' + f + ' -o ' + binname + '.exe'
+command_args = compiler_cmd + ' ' + f + ' -o ' + binname + '.exe'
+dat = "#!/usr/bin/env python\nimport os\nimport sys\nargs = sys.argv[1:]\nargs = (' ').join(args)\nprint('last compiled " + str(time_string) + "')\na = os.system('"+binname+".exe ' + args)\n"
+
+if ext == 'cpp' or ext == 'c':
+    print KYEL + command_args + KNRM
     a = os.system(command_args)
-    binfile = open(binname, "w");
-    binfile.write("#!/usr/bin/python\nimport sys,os\nargs=sys.argv;\nargs=args[1:];\nargs=(' ').join(args);\nprint('last compiled"+str(t)+"');\na=os.system('"+binname+".exe ' +args);\n");
-    binfile.close();
-elif ext == 'c':
-    command_args = compiler_cmd + " "+f+ " -o "+binname+".exe"
-    a=os.system(command_args)
-    binfile = open(binname, "w");
-    binfile.write("#!/usr/bin/python\nimport sys,os\nargs=sys.argv;\nargs=args[1:];\nargs=(' ').join(args);\nprint('"+str(t)+" last compiled');\na=os.system('"+binname+".exe ' +args);\n");
-    binfile.close();
+    open(binname, 'w').write(dat)
 elif ext == 'py':
-    srcfile = open(f)
-    src_contents = srcfile.read();
-    srcfile.close()
-    binfile = open((binname),'w')
-    binfile.write( "#!/usr/bin/python\nprint('"+time_string+" last compiled');\n"+src_contents)
-    binfile.close()
+    src_contents = open(f).read();
+    dat = "#!/usr/bin/python\nprint('last compiled " + time_string + "');\n" + src_contents
+    open((binname),'w').write(dat)
 else:
     error('extension not supported')
 
-print command_args
-f = open(bin_folder + '.' + cmdname + '.README', 'w')
-f.write('compile.py\n')
-f.write('time: ' + time_string + '\n');
-f.write('binary directory: ' + bin_folder + '\n')
-f.write('command arguments used: ' + command_args + '\n')
-f.close()
+printw(binname)
+
+readme_fn = bin_folder + '.' + cmdname + '.README'
+open(readme_fn, 'w').write('compile.py\ntime: ' + time_string + '\nbinary directory: ' + bin_folder + '\ncommand arguments used: ' + command_args + '\n')
 
 if(os.path.exists(binname)):
     os.system('chmod 755 ' + binname)
